flask-server/requirements.txt
flask
flask_bcrypt
flask_login
flask_jwt_extended
flask_sqlalchemy
python-dotenv
psycopg2-binary